+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.11 (2020-09-17)
 
+- updated the external link behaviour to skip in game notification
 - fixed an issue with too long stats at the market window (#814)
 
 ## 1.0.10 (2020-08-02)